:recycle: refactor(runtime): 优化 Handle 类型与 HeapValue 结构
- Handle 从 u32 改为 usize（平台原生指针大小）
- 新增 HeapValue::Struct 变体存储结构体字段
- TaskId 从 u32 改为 usize
- 简化 VM 执行器中 Handle 使用
- 修复浮点数取模除零检查
- 嵌入式执行器添加 44 个测试用例

+++ src/runtime/value/runtime_value.rs
@@ -105,7 +105,7 @@ pub struct FunctionId(pub u32);
 
 /// Task ID for async scheduling
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct TaskId(pub u32);
+pub struct TaskId(pub usize);
 
 /// Async state for lazy evaluation
 #[derive(Debug, Clone, PartialEq, Hash)]
